Split of Clients and Resources to separate modules
