+++ lib/three_scale_api/clients/account.rb
@@ -1,20 +1,21 @@
 # frozen_string_literal: true
 
-require 'three_scale_api/clients/default_helpers'
+require 'three_scale_api/clients/default'
+require 'three_scale_api/resources/account'
 require 'three_scale_api/clients/account_user'
 require 'three_scale_api/clients/application'
 
 module ThreeScaleApi
   module Clients
     # Accounts resource manager wrapper for default entity received by REST API
-    class AccountManager < DefaultStateManager
+    class AccountClient < DefaultClient
+      include DefaultStateClient
       # @api public
       # Creates instance of the Accounts resource manager
       #
       # @param [ThreeScaleQE::TestClient] http_client Instance of http client
       def initialize(http_client)
         super(http_client, entity_name: 'account', collection_name: 'accounts')
-        @resource_instance = Account
       end
 
       # Base path for the REST call
@@ -84,61 +85,6 @@ module ThreeScaleApi
       # @param [Fixnum] id Account ID
       def pending(id)
         set_state(id, 'make_pending')
-      end
-    end
-
-    # Default resource wrapper for any entity received by REST API
-    class Account < DefaultStateResource
-      # @api public
-      # Creates instance of the Service resource
-      #
-      # @param [ThreeScaleQE::TestClient] client Instance of the test client
-      # @param [AccountManager] manager Instance of the manager
-      # @param [Hash] entity Service Hash from API client
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
-      end
-
-      # @api public
-      # Sets plan for account
-      #
-      # @param [Fixnum] plan_id Plan ID
-      def set_plan(plan_id)
-        @manager.set_plan(@entity['id'], plan_id) if @manager.respond_to?(:set_plan)
-      end
-
-      # @api public
-      # Approves account
-      def approve
-        set_state('approve')
-      end
-
-      # @api public
-      # Reject account
-      def reject
-        set_state('reject')
-      end
-
-      # @api public
-      # Set pending for account
-      def pending
-        set_state('pending')
-      end
-
-      # @api public
-      # Gets Account Users Manager
-      #
-      # @return [AccountUsersManager] Account Users Manager
-      def users
-        manager_instance(AccountUserManager)
-      end
-
-      # @api public
-      # Gets  Application Manager
-      #
-      # @return [ApplicationManager] Account Users Manager
-      def applications
-        manager_instance(ApplicationManager)
       end
     end
   end
