+++ lib/three_scale_api/clients/account_user.rb
@@ -1,11 +1,15 @@
 # frozen_string_literal: true
 
-require 'three_scale_api/clients/default_helpers'
+require 'three_scale_api/clients/default'
+require 'three_scale_api/resources/account_user'
+
 
 module ThreeScaleApi
   module Clients
     # Account user resource manager wrapper for account user entity received by REST API
-    class AccountUserManager < DefaultUserManager
+    class AccountUserClient < DefaultClient
+      include DefaultUserClient
+
       attr_accessor :account
       # @api public
       # Creates instance of the Account user resource manager
@@ -13,8 +17,7 @@ module ThreeScaleApi
       # @param [ThreeScaleQE::TestClient] http_client Instance of http client
       # @param [Account] account Account entity
       def initialize(http_client, account)
-        super(http_client)
-        @resource_instance = AccountUser
+        super(http_client, entity_name: 'user', collection_name: 'users')
         @account = account
       end
 
@@ -23,21 +26,6 @@ module ThreeScaleApi
       # @return [String] Base URL for the REST call
       def base_path
         super.concat "/accounts/#{@account['id']}/users"
-      end
-    end
-
-    # Account user resource wrapper for account user received by REST API
-    class AccountUser < DefaultUserResource
-      attr_accessor :account
-      # @api public
-      # Creates instance of the User resource
-      #
-      # @param [ThreeScaleQE::TestClient] client Instance of the test client
-      # @param [AccountUserManager] manager Instance of the manager
-      # @param [Hash] entity Service Hash from API client
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
-        @account = manager.account
       end
     end
   end
