+++ lib/three_scale_api/clients/active_doc.rb
@@ -1,18 +1,19 @@
 # frozen_string_literal: true
 
 require 'three_scale_api/clients/default'
+require 'three_scale_api/resources/active_doc'
+
 
 module ThreeScaleApi
   module Clients
     # Active doc resource manager wrapper for the active doc entity received by the REST API
-    class ActiveDocManager < DefaultManager
+    class ActiveDocClient < DefaultClient
       # @api public
       # Creates instance of the Active doc resource manager
       #
       # @param [ThreeScaleQE::TestClient] http_client Instance of http client
       def initialize(http_client)
         super(http_client, entity_name: 'api_doc', collection_name: 'api_docs')
-        @resource_instance = ActiveDoc
       end
 
       # @api public
@@ -31,18 +32,6 @@ module ThreeScaleApi
       # @return [String] Base URL for the REST call
       def base_path
         super.concat '/active_docs'
-      end
-    end
-
-    # Active doc resource wrapper for the active doc entity received by REST API
-    class ActiveDoc < DefaultResource
-      # Creates instance of the Active doc resource
-      #
-      # @param [ThreeScaleQE::TestClient] client Instance of the test client
-      # @param [ActiveDocManager] manager Instance of the manager
-      # @param [Hash] entity Service Hash from API client
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
       end
     end
   end
