+++ lib/three_scale_api/clients/application_key.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
 require 'three_scale_api/clients/default'
+require 'three_scale_api/resources/application_key'
 
 module ThreeScaleApi
   module Clients
     # Application key resource manager wrapper for a application key entity received by the REST API
-    class ApplicationKeyManager < DefaultManager
+    class ApplicationKeyClient < DefaultClient
       attr_accessor :account, :application
       # @api public
       # Creates instance of the Application Key resource manager
@@ -14,7 +15,6 @@ module ThreeScaleApi
       # @param [Application] application Account entity
       def initialize(http_client, application)
         super(http_client, entity_name: 'key', collection_name: 'keys')
-        @resource_instance = ApplicationKey
         @account = application.account
         @application = application
       end
@@ -55,28 +55,6 @@ module ThreeScaleApi
         @log.info("Delete #{resource_name}: #{key}")
         @http_client.delete("#{base_path}/#{key}")
         true
-      end
-    end
-
-    # Application key resource wrapper for a application key entity received by the REST API
-    class ApplicationKey < DefaultResource
-      attr_accessor :account, :application
-      # @api public
-      # Creates instance of the Application key resource
-      #
-      # @param [ThreeScaleQE::TestClient] client Instance of the test client
-      # @param [ApplicationKeyManager] manager Instance of the manager
-      # @param [Hash] entity Application key Hash from API client
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
-        @account = manager.account
-        @application = manager.application
-      end
-
-      # @api public
-      # Deletes key resource
-      def delete
-        @manager.delete(entity['value']) if @manager.respond_to?(:delete)
       end
     end
   end
