+++ lib/three_scale_api/clients/application_plan_limit.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
 require 'three_scale_api/clients/default'
+require 'three_scale_api/resources/application_plan_limit'
 
 module ThreeScaleApi
   module Clients
     # Application plan limit resource manager wrapper for an application plan limit entity
     # received by the REST API
-    class ApplicationPlanLimitManager < DefaultManager
+    class ApplicationPlanLimitClient < DefaultClient
       attr_accessor :service, :application_plan, :metric
 
       # @api public
@@ -19,7 +20,6 @@ module ThreeScaleApi
         @service = app_plan.service
         @application_plan = app_plan
         @metric = metric
-        @resource_instance = ApplicationPlanLimit
       end
 
       # Base path for the REST call
@@ -35,24 +35,6 @@ module ThreeScaleApi
       # @param [Metric] metric Metric resource
       def set_metric(metric)
         @metric = metric
-      end
-    end
-
-    # Application plan limit resource wrapper for an application plan limit entity
-    class ApplicationPlanLimit < DefaultResource
-      attr_accessor :service, :metric
-
-      # @api public
-      # Construct an application plan limit resource
-      #
-      # @param [ThreeScaleApi::HttpClient] client Instance of test client
-      # @param [ApplicationPlanLimitManager] manager Instance of the manager
-      # @param [Hash] entity Entity Hash from API client of the application plan limit
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
-        @service = manager.service
-        @metric = manager.metric
-        @application_plan = manager.application_plan
       end
     end
   end
