+++ lib/three_scale_api/clients/method.rb
@@ -2,12 +2,13 @@
 
 require 'three_scale_api/clients/default'
 require 'three_scale_api/clients/application_plan_limit'
-require 'three_scale_api/clients/method'
+require 'three_scale_api/resources/method'
+
 
 module ThreeScaleApi
   module Clients
     # Method resource manager wrapper for the method entity received by REST API
-    class MethodManager < DefaultManager
+    class MethodClient < DefaultClient
       attr_accessor :service, :metric
 
       # @api public
@@ -19,7 +20,6 @@ module ThreeScaleApi
         super(http_client, entity_name: 'method', collection_name: 'methods')
         @service = metric.service
         @metric = metric
-        @resource_instance = Method
       end
 
       # Base path for the REST call
@@ -27,31 +27,6 @@ module ThreeScaleApi
       # @return [String] Base URL for the REST call
       def base_path
         super.concat "/services/#{@service['id']}/metrics/#{@metric['id']}/methods"
-      end
-    end
-
-    # @api public
-    # Method resource wrapper for the metric entity received by the REST API
-    class Method < DefaultResource
-      attr_accessor :service, :metric
-
-      # Construct the metric resource
-      #
-      # @param [ThreeScaleApi::HttpClient] client Instance of http client
-      # @param [ThreeScaleApi::Clients::MethodManager] manager Method manager
-      # @param [Hash] entity Entity Hash from API client of the metric
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
-        @service = manager.service
-        @metric = manager.metric
-      end
-
-      # @api public
-      # Gets application plan limits
-      #
-      # @return [ApplicationPlanLimitManager] Instance of the Application plan limits manager
-      def application_plan_limits(app_plan)
-        ApplicationPlanLimitManager.new(@http_client, app_plan, metric: self)
       end
     end
   end
