+++ lib/three_scale_api/clients/provider.rb
@@ -1,19 +1,22 @@
 # frozen_string_literal: true
 
 require 'three_scale_api/tools'
-require 'three_scale_api/clients/default_helpers'
+require 'three_scale_api/clients/default'
+require 'three_scale_api/resources/provider'
+
 
 module ThreeScaleApi
   module Clients
     # Provider resource manager wrapper for the provider entity received by the REST API
-    class ProviderManager < DefaultUserManager
+    class ProviderClient < DefaultClient
+      include DefaultUserClient
+
       # @api public
       # Creates instance of the Provider resource manager
       #
       # @param [ThreeScaleQE::TestClient] http_client Instance of http client
       def initialize(http_client)
-        super(http_client)
-        @resource_instance = Provider
+        super(http_client, entity_name: 'user', collection_name: 'users')
       end
 
       # Base path for the REST call
@@ -21,19 +24,6 @@ module ThreeScaleApi
       # @return [String] Base URL for the REST call
       def base_path
         super.concat '/users'
-      end
-    end
-
-    # Provider resource wrapper for the provider entity received by REST API
-    class Provider < DefaultUserResource
-      # @api public
-      # Creates instance of the Provider resource
-      #
-      # @param [ThreeScaleQE::TestClient] client Instance of the test client
-      # @param [ProviderManager] manager Instance of the provider manager
-      # @param [Hash] entity Provider Hash from API client
-      def initialize(client, manager, entity)
-        super(client, manager, entity)
       end
     end
   end
