+++ lib/three_scale_api/resources/settings.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+require 'three_scale_api/resources/default'
+
+module ThreeScaleApi
+  module Resources
+    # Service resource wrapper for the service entity received by REST API
+    class Settings < DefaultResource
+      # @api public
+      # Creates instance of the Service resource
+      #
+      # @param [ThreeScaleQE::TestClient] client Instance of the test client
+      # @param [ServicesManager] manager Instance of the manager
+      # @param [Hash] entity Service Hash from API client
+      def initialize(client, manager, entity)
+        super(client, manager, entity)
+      end
+    end
+  end
+end
